Backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
motor>=3.2.0
pydantic>=2.3.0
pydantic>=2.0.0,<2.3.0
email-validator>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
yagmail>=0.15.293
certifi>=2023.5.7